Refactor atom package structure
