+++ crdt/atom/doc.go
@@ -1,0 +1,21 @@
+/*
+	This package implements the atom for the causal tree.
+
+	Description
+
+	This package defines three types: AtomID, AtomValue and Atom.
+	And its methods. The atom type is the basic unit of the causal tree.
+
+	AtomID
+
+	AtomID is the unique identifier of an atom.
+
+	AtomValue
+
+	AtomValue is a tree operation.
+
+	Atom
+
+	Atom represents an atomic operation within a replicated tree.
+*/
+package atom
